requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
dateparser==1.2.2
fonttools==4.59.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.5
multidict==6.6.3
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
pycryptodome==3.23.0
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
regex==2025.7.34
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
ta==0.11.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
